WIP: делит редьюсер на части
