+++ src/reducer/show-films/show-films.js
@@ -1,0 +1,58 @@
+import {extend} from '../../utils.js';
+import {ALL_GENRES} from '../../const.js';
+
+const initialState = {
+  currentGenre: ALL_GENRES,
+};
+
+const ActionType = {
+  CHOOSE_GENRE: `CHOOSE_GENRE`,
+  GET_ALL_FILMS: `GET_ALL_FILMS`,
+  GET_FILMS_BY_GENRE: `GET_FILMS_BY_GENRE`,
+};
+
+const ActionCreator = {
+  chooseGenre: (genre) => ({
+    type: ActionType.CHOOSE_GENRE,
+    payload: genre,
+  }),
+
+  getAllFilms: (films) => ({
+    type: ActionType.GET_ALL_FILMS,
+    payload: films,
+  }),
+
+  getFilmsByGenre: (selectedGenre, films) => {
+    const filmsByGenre = films
+        .filter((film) => film.genre === selectedGenre);
+
+    return {
+      type: ActionType.GET_FILMS_BY_GENRE,
+      payload: filmsByGenre,
+    };
+  },
+};
+
+const reducer = (state = initialState, action) => {
+  switch (action.type) {
+    case ActionType.CHOOSE_GENRE:
+      return extend(state, {
+        currentGenre: action.payload,
+      });
+
+    case ActionType.GET_ALL_FILMS:
+      return extend(state, {
+        filmsByGenre: action.payload
+      });
+
+    case ActionType.GET_FILMS_BY_GENRE:
+      return extend(state, {
+        filmsByGenre: action.payload,
+      });
+
+    default:
+      return state;
+  }
+};
+
+export {ActionCreator, ActionType, reducer};
